Neovim: Add buganizer and cider ML nvim-cmp sources.
This really does not belong in my public dotfiles, but since it can't be
pulled out of the nvim-cmp setup, I've decided to include it. It's
commented out so that on systems where these sources are not available,
this won't cause any problems.

The names of these sources being included in this public repo does not
present any security concerns.